Refactor by creating `TasksListElement` class
